Move adding weather data to cleaning data
